AssetManagement WiP #CHANGE: Moved all AssetManagement classes into their own Duality.Editor.AssetManagement namespace.
