+++ DualityEditorPlugins/EditorBase/Importers/PixmapAssetImporter.cs
@@ -8,6 +8,7 @@ using Duality.IO;
 using Duality.Drawing;
 using Duality.Resources;
 using Duality.Editor;
+using Duality.Editor.AssetManagement;
 
 namespace Duality.Editor.Plugins.Base
 {
